backend/requirements.txt
flask
flask-cors
yfinance
pandas
numpy
psycopg2-binary
python-dotenv
gunicorn
scipy
scikit-learn
scikit-learn
